server/src/services/technicalAnalysis.ts
import { SMA, RSI, MACD, BollingerBands } from 'technicalindicators';
import { MarketData, MarketStructure, TimeFrame, TrendDirection, TrendStrength } from '../types/market';

export enum IndicatorType {
  SMA = 'sma',
  RSI = 'rsi',
  MACD = 'macd',
  BB = 'bb',
  VWAP = 'vwap'
}

interface IndicatorConfig {
  type: IndicatorType;
  period?: number;
  shortPeriod?: number;
  longPeriod?: number;
  signalPeriod?: number;
  standardDeviation?: number;
}

export class TechnicalAnalysis {
  private readonly defaultPeriod = 14;
  private readonly defaultShortPeriod = 12;
  private readonly defaultLongPeriod = 26;
  private readonly defaultSignalPeriod = 9;
  private readonly defaultStdDev = 2;

  public async analyze(
    symbol: string,
    timeframe: TimeFrame,
    indicators: Array<{ type: IndicatorType }> = []
  ): Promise<MarketStructure> {
    try {
      const marketData = await this.fetchMarketData(symbol, timeframe);
      if (!marketData || marketData.length === 0) {
        throw new Error('Insufficient market data for analysis');
      }

      const trend = this.analyzeTrend(marketData);
      const volatility = {
        value: this.calculateVolatility(marketData),
        timestamp: marketData[marketData.length - 1].timestamp
      };
      const momentum = this.calculateMomentum(marketData);

      return {
        trend,
        volatility,
        momentum: momentum.value
      };
    } catch (error) {
      throw error;
    }
  }

  private async fetchMarketData(symbol: string, timeframe: TimeFrame): Promise<MarketData[]> {
    // Implementation depends on your data source
    // This should return an array of MarketData objects
    throw new Error('Method not implemented');
  }

  private calculateVolatility(marketData: MarketData[]): number {
    if (marketData.length <= 1) {
      return 0;
    }

    const returns = marketData.map(data => data.close);
    const mean = returns.reduce((sum, val) => sum + val, 0) / returns.length;
    const squaredDiffs = returns.map(val => Math.pow(val - mean, 2));
    return Math.sqrt(squaredDiffs.reduce((sum, val) => sum + val, 0) / (returns.length - 1));
  }

  private analyzeTrend(marketData: MarketData[]): { direction: TrendDirection; strength: TrendStrength; startTime?: Date; endTime?: Date } {
    if (marketData.length <= 1) {
      return {
        direction: TrendDirection.Sideways,
        strength: TrendStrength.VeryWeak,
        startTime: marketData[0].timestamp,
        endTime: marketData[0].timestamp
      };
    }

    const prices = marketData.map(data => data.close);
    const firstPrice = prices[0];
    const lastPrice = prices[prices.length - 1];
    const priceChange = lastPrice - firstPrice;

    // Calculate trend strength based on price movement consistency
    let upMoves = 0;
    let downMoves = 0;
    for (let i = 1; i < prices.length; i++) {
      const diff = prices[i] - prices[i - 1];
      if (diff > 0) upMoves++;
      else if (diff < 0) downMoves++;
    }

    const consistency = Math.max(upMoves, downMoves) / (prices.length - 1);
    let strength: TrendStrength;

    const threshold = 0.0001; // Minimum price change to consider trend
    const direction = Math.abs(priceChange) < threshold ? 
      TrendDirection.Sideways : 
      (priceChange > 0 ? TrendDirection.Up : TrendDirection.Down);

    // For sideways trend, use a different strength calculation
    if (direction === TrendDirection.Sideways) {
      strength = TrendStrength.Moderate; // Default to moderate strength for sideways
    } else {
      // For up/down trends, use consistency
      if (consistency > 0.8) strength = TrendStrength.VeryStrong;
      else if (consistency > 0.6) strength = TrendStrength.Strong;
      else if (consistency > 0.4) strength = TrendStrength.Moderate;
      else if (consistency > 0.2) strength = TrendStrength.Weak;
      else strength = TrendStrength.VeryWeak;
    }

    return {
      direction,
      strength,
      startTime: marketData[0].timestamp,
      endTime: marketData[marketData.length - 1].timestamp
    };
  }

  private calculateMomentum(marketData: MarketData[]): { value: number; timestamp: Date } {
    if (marketData.length <= 1) {
      return {
        value: 0,
        timestamp: marketData[0].timestamp
      };
    }

    const prices = marketData.map(data => data.close);
    const roc = ((prices[prices.length - 1] - prices[0]) / prices[0]) * 100;

    return {
      value: roc,
      timestamp: marketData[marketData.length - 1].timestamp
    };
  }

  private calculateSMA(values: number[], period: number): number[] {
    return SMA.calculate({ period, values });
  }

  private calculateRSI(values: number[]): number[] {
    return RSI.calculate({ period: this.defaultPeriod, values });
  }

  private calculateMACD(values: number[]): number[] {
    const macdValues = MACD.calculate({
      values,
      fastPeriod: this.defaultShortPeriod,
      slowPeriod: this.defaultLongPeriod,
      signalPeriod: this.defaultSignalPeriod,
      SimpleMAOscillator: false,
      SimpleMASignal: false
    });
    
    return macdValues.map(r => r.MACD);
  }

  private calculateBollingerBands(values: number[]): {
    upper: number[];
    middle: number[];
    lower: number[];
  } {
    const bb = BollingerBands.calculate({
      period: this.defaultPeriod,
      values,
      stdDev: this.defaultStdDev
    });
    
    return {
      upper: bb.map(b => b.upper),
      middle: bb.map(b => b.middle),
      lower: bb.map(b => b.lower)
    };
  }

  public calculateVWAP(data: MarketData[]): number {
    let cumulativeTPV = 0;
    let cumulativeVolume = 0;

    for (const candle of data) {
      const typicalPrice = (candle.high + candle.low + candle.close) / 3;
      const volume = candle.volume;
      
      cumulativeTPV += typicalPrice * volume;
      cumulativeVolume += volume;
    }

    return cumulativeTPV / cumulativeVolume;
  }
}
